integration tests - rework deposit
